+++ index.js
@@ -1,7 +1,9 @@
 var createLoguxCreator = require('./create-logux-creator')
+var loguxStoreEnhancer = require('./logux-store-enhancer')
 var loguxReducer = require('./logux-reducer')
 
 module.exports = {
-  createLoguxCreator: createLoguxCreator
+  createLoguxCreator: createLoguxCreator,
+  loguxStoreEnhancer: loguxStoreEnhancer,
   loguxReducer: loguxReducer
 }
